switch from hook to epic quipu apy loading
