+++ src/store/wallet/wallet-epics.ts
@@ -18,6 +18,7 @@ import { showErrorToast } from '../../toast/toast.utils';
 import { getTokenSlug } from '../../token/utils/token.utils';
 import { groupActivitiesByHash } from '../../utils/activity.utils';
 import { mapOperationsToActivities } from '../../utils/operation.utils';
+import { loadQuipuApy$ } from '../../utils/quipu-apy.util';
 import { createReadOnlyTezosToolkit, CURRENT_NETWORK_ID } from '../../utils/rpc/tezos-toolkit.utils';
 import { loadAssetsBalances$, loadTezosBalance$, loadTokensWithBalance$ } from '../../utils/token-balance.utils';
 import { loadTokenMetadata$, loadTokensWithBalanceMetadata$ } from '../../utils/token-metadata.utils';
@@ -30,6 +31,7 @@ import { navigateAction } from '../root-state.actions';
 import {
   addTokenMetadataAction,
   loadActivityGroupsActions,
+  loadQuipuApyActions,
   loadTezosBalanceActions,
   loadTokenBalancesActions,
   loadTokenMetadataActions,
@@ -42,8 +44,20 @@ const updateDataActions = () => [
   loadTezosBalanceActions.submit(),
   loadTokenBalancesActions.submit(),
   loadActivityGroupsActions.submit(),
+  loadQuipuApyActions.submit(),
   loadSelectedBakerActions.submit()
 ];
+
+const loadQuipuApyEpic = (action$: Observable<Action>) =>
+  action$.pipe(
+    ofType(loadQuipuApyActions.submit),
+    switchMap(() =>
+      loadQuipuApy$().pipe(
+        map(apy => loadQuipuApyActions.success(apy)),
+        catchError(err => of(loadQuipuApyActions.fail(err.message)))
+      )
+    )
+  );
 
 const loadTokenBalancesEpic = (action$: Observable<Action>, state$: Observable<RootState>) =>
   action$.pipe(
@@ -188,6 +202,7 @@ const addTokenMetadataEpic = (action$: Observable<Action>) =>
 
 export const walletEpics = combineEpics(
   loadTezosBalanceEpic,
+  loadQuipuApyEpic,
   loadTokenBalancesEpic,
   loadTokenSuggestionEpic,
   loadTokenMetadataEpic,
