+++ src/store/wallet/wallet-state.mock.ts
@@ -2,6 +2,7 @@ import { mockHdAccount } from '../../interfaces/account.interface.mock';
 import { mockWalletAccountState } from '../../interfaces/wallet-account-state.interface.mock';
 import { mockFA1_2TokenMetadata, mockFA2TokenMetadata } from '../../token/interfaces/token-metadata.interface.mock';
 import { getTokenSlug } from '../../token/utils/token.utils';
+import { QUIPU_DEFAULT_PERCENTAGE } from '../../utils/quipu-apy.util';
 import { createEntity } from '../create-entity';
 import { WalletState } from './wallet-state';
 
@@ -13,5 +14,6 @@ export const mockWalletState: WalletState = {
     [getTokenSlug(mockFA2TokenMetadata)]: mockFA2TokenMetadata
   },
   addTokenSuggestion: createEntity(mockFA1_2TokenMetadata),
-  isShownDomainName: false
+  isShownDomainName: false,
+  quipuApy: QUIPU_DEFAULT_PERCENTAGE
 };
